[SofaKernel] Remove some dependencies from SofaHelper to SofaCore
