+++ SofaKernel/modules/SofaHelper/src/sofa/helper/AdvancedTimer.cpp
@@ -25,6 +25,7 @@
 #include <sofa/helper/system/thread/CTime.h>
 #include <sofa/helper/vector.h>
 #include <sofa/helper/map.h>
+#include <iomanip>
 #include "../../extlibs/json/json.h"
 
 
@@ -36,7 +37,6 @@
 
 #define DEFAULT_INTERVAL 100
 
-using namespace sofa::core::objectmodel;
 using json = sofa::helper::json;
