+++ SofaKernel/modules/SofaHelper/src/sofa/helper/io/MeshGmsh.h
@@ -49,9 +49,9 @@ protected:
 
     bool readGmsh(std::ifstream &file, const unsigned int gmshFormat);
 
-    void addInGroup(helper::vector< sofa::core::loader::PrimitiveGroup>& group, int tag, std::size_t eid);
+    void addInGroup(helper::vector< sofa::helper::types::PrimitiveGroup>& group, int tag, std::size_t eid);
 
-    void normalizeGroup(helper::vector< sofa::core::loader::PrimitiveGroup>& group);
+    void normalizeGroup(helper::vector< sofa::helper::types::PrimitiveGroup>& group);
 };
 
 } // namespace io
