+++ SofaKernel/modules/SofaHelper/src/sofa/helper/types/PrimitiveGroup.h
@@ -19,4 +19,26 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#error This file will be removed after sofa 21.06 release. Update your code by replacing #include<sofa/defaulttype/Vec3Types.h> with #include<sofa/defaulttype/VecTypes.h>
+#pragma once
+#include <string>
+
+namespace sofa::helper::types
+{
+
+class PrimitiveGroup
+{
+public:
+    int p0, nbp;
+    std::string materialName;
+    std::string groupName;
+    int materialId;
+
+    friend std::ostream& operator << (std::ostream& out, const PrimitiveGroup &g);
+    friend std::istream& operator >> (std::istream& in, PrimitiveGroup &g);
+    bool operator <(const PrimitiveGroup& p) const;
+
+    PrimitiveGroup();
+    PrimitiveGroup(int p0, int nbp, std::string materialName, std::string groupName, int materialId);
+};
+
+} // namespace sofa
